ENH: remove the single filament limitation of skipping object
We planed to skip uncessary filament changing when skip
object. However, this caused color mixing.

If we try to fix the color mixing issue, the flush gcode
should be generated by printer itself. This is not a short
time job.

So currently we never skip changing filament when skip
object. This will not save much time and flushing. But better
than none.

Signed-off-by: salt.wei <salt.wei@bambulab.com>
Change-Id: I6046cd2765cf68058712b4256402acc0be771144