NEW:add measure gizmo new files
Jira: STUDIO-6166
most of the code come from orca
The original code is from PrusaSlicer and referring to the implementation of OrcaSlicer. Thanks for PrusaSlicer and OrcaSlicer , thanks for Lukas Matena and enricoturri1966 and so on.
commit f72d42f920334a05b0efa01f18eb4cc066809c00
Author: enricoturri1966 <enricoturri@seznam.cz>
Date:   Tue Oct 31 11:43:04 2023 +0800

    Measure: Initial porting of Measure Gizmo

diff --git a/resources/images/copy_menu.svg b/resources/images/copy_menu.svg
new file mode 100644

Change-Id: I81afa52165f14c34e0c63f2c40c953a04d76fc8b